trainer/training/requirements.txt
tensorflow==2.16.1
tensorflow-model-optimization
scikit-learn
opencv-python-headless
pyyaml
numpy
matplotlib
seaborn
git+https://github.com/sebastian-sz/efficientnet-lite-keras@main
